% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{icc.multimeasure}
\alias{icc.multimeasure}
\title{ICC calculation for a}
\usage{
\method{icc}{multimeasure}(object, method = "fisher", ...)
}
\arguments{
\item{object}{multimeasure object}

\item{method}{"fisher" "randeff" or "randeff2"}
}
\description{
ICC calculation for a
}
